fix(backend): Use the project root dir to find static file resources
